refactor(parser): extract line and raw-data helpers
